requirements.txt
fastapi==0.62.0
fastapi-utils==0.2.1
uvicorn==0.12.3
gunicorn==20.0.4
uvloop==0.14.0
httptools==0.1.1
requests==2.24.0
PyJWT==1.4.2
python-multipart==0.0.5
python-json-logger==0.1.11
aiofiles==0.6.0
redis==3.5.3
python-dotenv==0.19.1
pydantic==1.8.2
elasticsearch==7.15.2
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-asyncpg==0.26b1
opentelemetry-instrumentation-elasticsearch==0.26b1
